Add guide and sample code for working with GUIDs
A new tutorial has been added to the documentation explaining how to work with IDs, specifically GUIDs. This improvement includes a guide on how to customize IDs generation to be optimized for SQL server. Additionally, new sample code is provided to showcase how to generate sequential GUIDs using a factory pattern and how to work with these value objects representing IDs. This information is valuable for developers working with Vogen who want better performance with SQL Server GUID columns.