Update Vogen version and expand FAQ section
The Vogen package reference in the ConsumerTests.csproj file was updated to version 4.0.13 from 4.0.4. Additionally, the FAQ.md file under the docs folder was amended to include information on how to use value objects as parameters in Blazor pages and components. The layout of the tutorial "Working-with-Guids.md" in hi.tree was also slightly rearranged.