Update package version and documentation content
The Installation instruction now uses a variable for the package version, improving the ease of maintenance when new releases are made. Detailed information on shared types in the package has been added, and the 'Working-with-Guids' document has been renamed to 'Working-with-IDs'. Other minor text updates have also been applied throughout the documentation.