Fix invalid strong name on assembly
The issue here was that although you have the private key with which to sign, the build wasn't using it.

`PublicSign=true` tells the build to _not_ sign the assembly, but merely give it the strong name. This is fine for .NET but breaks in certain scenarios on .NET Framework.